requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
anaconda-anon-usage @ file:///C:/b/abs_1az9wuczyu/croot/anaconda-anon-usage_1749054795139/work
annotated-types @ file:///C:/b/abs_0dmaoyhhj3/croot/annotated-types_1709542968311/work
archspec @ file:///croot/archspec_1709217642129/work
asttokens==3.0.0
blinker==1.9.0
boltons @ file:///C:/b/abs_e2_iokhxbp/croot/boltons_1751383740243/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/miniconda3/conda-bld/certifi_1759786080784/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
comm==0.2.2
conda @ file:///C:/b/abs_78fae1xfc9/croot/conda_1749171262913/work/conda-src
conda-anaconda-telemetry @ file:///C:/b/abs_d4qxn0oeix/croot/conda-anaconda-telemetry_1751062015410/work
conda-anaconda-tos @ file:///C:/b/abs_4csynbis2x/croot/conda-anaconda-tos_1752801584932/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_2d41dz5u60/croot/conda-package-streaming_1751307521903/work
contourpy==1.3.0
crcmod==1.7
cryptography @ file:///C:/b/abs_5eg9xpv31w/croot/cryptography_1748977441610/work
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.14.0
Flask==2.3.0
fonttools==4.60.1
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
fsspec==2025.9.0
ftfy==6.3.1
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
gunicorn==23.0.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata==8.7.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jmespath==0.10.0
joblib==1.5.2
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
libmambapy @ file:///C:/b/abs_627vsv8bhu/croot/mamba-split_1734469608328/work/libmambapy
Markdown==3.9
markdown-it-py @ file:///C:/b/abs_a5bfngz6fu/croot/markdown-it-py_1684279915556/work
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.7.0
matplotlib-inline==0.1.7
mdurl @ file:///C:/Windows/TEMP/abs_3197pzpjbi/croots/recipe/mdurl_1659716032440/work
menuinst @ file:///C:/b/abs_a0aibbe6s6/croot/menuinst_1750861726761/work
mkl==2021.4.0
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mmcv==2.1.0
mmengine==0.10.7
-e git+https://github.com/open-mmlab/mmsegmentation.git@b040e147adfa027bbc071b624bedf0ae84dfc922#egg=mmsegmentation
model-index==0.1.11
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==1.24.0
opencv-python==4.8.0.74
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.3
ordered-set==4.1.0
oss2==2.17.0
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.3.3
parso==0.8.4
Pillow==10.0.0
platformdirs==4.3.8
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
prettytable==3.16.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.23.0
pydantic @ file:///C:/b/abs_d1rp6fqxgk/croot/pydantic_1750768285689/work
pydantic_core @ file:///C:/b/abs_d0xs0z3erk/croot/pydantic-core_1750754747982/work
Pygments==2.19.1
pyparsing==3.2.5
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
pytz==2023.4
pywin32==310
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq==26.4.0
regex==2025.11.3
requests==2.28.2
rich==13.4.2
ruamel.yaml @ file:///C:/b/abs_0dn1nhdq75/croot/ruamel.yaml_1745960327187/work
ruamel.yaml.clib @ file:///C:/b/abs_475flulhy_/croot/ruamel.yaml.clib_1745937166245/work
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
stack-data==0.6.3
sympy @ file:///C:/b/abs_f1v7nwa6vp/croot/sympy_1756713483242/work
tabulate==0.9.0
tbb==2021.11.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
typing-inspection @ file:///C:/b/abs_261c5jt6kx/croot/typing-inspection_1746023491500/work
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==2.3.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
yapf==0.43.0
zipp==3.21.0
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
